refactor: Move PyQt6-specific code out of shared UI directory
**Problem**: CI pipeline failing with PyQt6 import errors because PyQt6-specific code was in the shared UI directory (openhcs/ui/shared/), causing PyQt6 imports during normal operation.

**Root Cause**: Architectural violation - shared code had framework-specific dependencies, breaking separation of concerns.

**Solution**: Proper architectural restructuring:
- Move pyqt6_widget_strategies.py → openhcs/pyqt_gui/widgets/shared/widget_strategies.py
- Update widget_creation_registry.py to use lazy imports with graceful fallback
- Update parameter_form_manager.py to import from correct location
- Revert CI workflows to only install [dev] extra (no PyQt6 dependencies)

**Technical Benefits**:
- ✅ Clean separation: Shared code has no framework dependencies
- ✅ Lazy imports: PyQt6 code only loaded when GUI is actually used
- ✅ Graceful fallback: Clear error messages when PyQt6 unavailable
- ✅ CI compatibility: Headless testing without PyQt6 requirements
- ✅ Maintainable: No conditional imports or placeholder classes

**Architecture**: Framework-specific code stays in framework directories, shared code remains truly framework-agnostic.
