+++ .github/workflows/coverage-pages.yml
@@ -32,7 +32,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e ".[dev,gui,dev-gui]"
+          pip install -e ".[dev]"
 
       - name: Run integration tests with coverage
         continue-on-error: true  # Continue workflow even if tests fail
