+++ openhcs/pyqt_gui/widgets/shared/parameter_form_manager.py
@@ -20,7 +20,7 @@ from openhcs.ui.shared.parameter_form_constants import CONSTANTS
 
 from openhcs.ui.shared.widget_creation_registry import create_pyqt6_registry
 from openhcs.ui.shared.ui_utils import format_param_name, format_field_id, format_reset_button_id
-from openhcs.ui.shared.pyqt6_widget_strategies import PyQt6WidgetEnhancer
+from .widget_strategies import PyQt6WidgetEnhancer
 
 # Import PyQt-specific components
 from .clickable_help_components import GroupBoxWithHelp, LabelWithHelp
